Rename kafkacat to kcat
Apache Software Foundation (ASF) trademark policies prohibits the use
of an Apache project name as part of a non-Apache project name.
In this case it was the use of `kafka` in `kafkacat` that had to be changed.

Apart from the name change there are no functional differences.
The now deprecated kafkacat.conf and KAFKACAT_CONF file and env vars
will continue to work for some time, but users should change to
kcat.conf and KCAT_CONF.